Flatten DirectMonotonicReader memory layout
We can do a couple of optimizations here directly and this sets the
stage for additional high-impact optimizations as well.
Quite often the DirectMonotonicReader contain only a single block
that can be unwrapped, or values were linear and as a result none of the
blocks contain any readers and similar. This implements a few
straight-forward optimizations of this kind.